% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_inertia_ahc.R
\name{clust_centroid}
\alias{clust_centroid}
\title{Centroid of a cluster}
\usage{
clust_centroid(i, data, cluster)
}
\arguments{
\item{i}{an integer that represents the cluster number}

\item{data}{a R data frame (all columns are required to be numeric types)}

\item{cluster}{a character. This refers to the column name of the data frame representing the clusters}
}
\value{
a vector of coordinates for the centroid of the cluster i
}
\description{
This function allows to compute the centroid of a cluster in a R data frame
}
\references{
Link to the author's github package repository:
\url{https://www.github.com/Redcart/helda}
}
\author{
Simon CORDE
}
\keyword{centroids}
\keyword{data}
\keyword{frame}
