% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_centroid.R
\name{cluster_centroid}
\alias{cluster_centroid}
\title{Centroid of a cluster}
\usage{
cluster_centroid(i, data, cluster)
}
\arguments{
\item{i}{an integer that represents the cluster number.}

\item{data}{a R data frame (all columns are required to be numeric types).}

\item{cluster}{a character. This refers to the column name of the data frame representing the clusters.}
}
\value{
a vector of coordinates of the centroid of the cluster i.
}
\description{
This function allows to compute the centroid of a cluster in a R data frame
}
\references{
Link to the author's github package repository:
\url{https://www.github.com/Redcart/helda}
}
\author{
Simon CORDE
}
\keyword{centroids}
\keyword{data}
\keyword{frame}
