% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_inertia_ahc.R
\name{compute_inertia}
\alias{compute_inertia}
\title{Inertia of data frame}
\usage{
compute_inertia(data)
}
\arguments{
\item{data}{R data frame (all columns are required to be numeric types)}
}
\description{
This function allows to compute the inertia in a R data frame
}
\references{
Link to the author's github repository:
\url{https://www.github.com/Redcart}
}
\author{
Simon CORDE
}
\keyword{data}
\keyword{frame}
\keyword{inertia}
