% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_inertia_ahc.R
\name{compute_inertia_ahc}
\alias{compute_inertia_ahc}
\title{Intragroup inertia for choosing the optimal number of clusters in Agglomerative Clustering}
\usage{
compute_inertia_ahc(data, method = "ward.D", max_clusters = 10)
}
\arguments{
\item{data}{R data frame (all columns are required to be numeric types)}

\item{method}{character that specifies the method on which the agglomerative is built upon}

\item{max_clusters}{maximal number of clusters for which we intend to compute intra group inertia}
}
\value{
vector of length max_clusters containing the inter group inertia for agglomerative
clustering with 1 cluster to max_clusters
}
\description{
This function allows to compute the inter group inertia from agglomerative clustering
for different number of clusters
}
\examples{
data = iris[c(1:4)]
results <- compute_inertia_ahc(data=data, max_clusters=15)
}
\references{
Link to the author's github repository:
\url{https://www.github.com/Redcart}
}
\author{
Simon CORDE
}
\keyword{agglomerative}
\keyword{clustering}
\keyword{data}
\keyword{frame}
\keyword{inertia}
