% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_formula_test.R
\docType{data}
\name{create_formula_test}
\alias{create_formula_test}
\title{String representing a R formula for unit test of create_formula function}
\format{Formula with features of Iris data set}
\usage{
create_formula_test
}
\description{
String representing a R formula for unit test of create_formula function
}
\examples{
data(create_formula_test)
}
\keyword{datasets}
