% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_test.R
\docType{data}
\name{formula_test}
\alias{formula_test}
\title{Airline names.}
\format{Data frame with columns
\describe{
\item{carrier}{Two letter abbreviation.}
\item{name}{Full name.}
}}
\source{
<https://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236>
}
\usage{
formula_test
}
\description{
Look up airline names from their carrier codes.
}
\examples{
data(formula_test)
}
\keyword{datasets}
