% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_procedure.R
\name{kmeans_procedure}
\alias{kmeans_procedure}
\title{Kmeans procedure}
\usage{
kmeans_procedure(
  data,
  columns,
  threshold_min,
  threshold_max,
  verbose = FALSE,
  seed = 42
)
}
\arguments{
\item{data}{R data frame}

\item{columns}{columns of the data frame on which we perform the kmeans algorithm}

\item{threshold_min}{minimum size for cluster}

\item{threshold_max}{maximum size fo cluster}

\item{verbose}{if TRUE print the current state of the procedure (by default FALSE)}

\item{seed}{seed for the random call (if we want the output to be reproducible)}
}
\description{
This function allows to perform kmeans clustering with some desired output sizes for clusters
}
\details{
The five columns diplay the following information:
\itemize{
 \item the first one shows the different categories of the variable on which we want to perform proc_freq
 \item then we have frequency, percentage, cumulative frequency and cumulative percentage
 }
}
\examples{
data <- iris
str(data)
result_array <- proc_freq(data$Species)
}
\references{
Link to the author's github repository:
\url{https://www.github.com/Redcart}
}
\author{
Simon CORDE
}
\keyword{cluster}
\keyword{kmeans}
\keyword{sizes}
