% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift_curve.R
\name{lift_curve}
\alias{lift_curve}
\title{Lift curve graph}
\usage{
lift_curve(predictions, true_labels, positive_label)
}
\arguments{
\item{predictions}{vector of predictions. These are generally the result of a machine learning model.
The predictions must be probabilities (a real number between 0 and 1).}

\item{true_labels}{vector of true labels.}

\item{positive_label}{string that specify the positive label (Y=1) in the `true_labels`}
}
\value{
plot the lift curve
}
\description{
This function allows to draw a lift curve in a ggplot style for binary classification model
}
\examples{
data <- iris
str(data)
result_array <- proc_freq(data$Species)
View(result_array)

}
\references{
Link to the author's github repository:
\url{https://www.github.com/Redcart}
}
\author{
Simon CORDE
}
\keyword{SAS}
\keyword{categorical}
\keyword{frequency}
\keyword{proc_freq}
\keyword{table}
\keyword{variable}
