% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift_effect.R
\name{lift_effect}
\alias{lift_effect}
\title{Lift effect curve}
\usage{
lift_effect(predictions, true_labels, positive_label)
}
\arguments{
\item{predictions}{vector of predictions. These are generally the result of a machine learning model.
The predictions must be probabilities (a real number between 0 and 1).}

\item{true_labels}{vector of true labels.}

\item{positive_label}{string that specify the positive label (Y=1) in the `true_labels`}
}
\value{
a ggplot object containing the lift effect
}
\description{
This function allows to draw the lift effect on a graph for binary classification model
}
\references{
Link to the author's github repository:
\url{https://www.github.com/Redcart}
}
\author{
Simon CORDE
}
\keyword{classification}
\keyword{curve}
\keyword{learning}
\keyword{lift}
\keyword{machine}
