% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_freq.R
\name{proc_freq}
\alias{proc_freq}
\title{This function permits to reproduce the output of the SAS proc freq}
\usage{
proc_freq(variable, digits = 4)
}
\arguments{
\item{variable}{vector on which we want to apply the function}

\item{digits:}{integer that specifies the number of decimals we want to keep in the rounded figures}
}
\description{
This function permits to reproduce the output of the SAS proc freq
}
\examples{
proc_freq(variable, digits = 4)

}
