% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_freq_test.R
\docType{data}
\name{proc_freq_test}
\alias{proc_freq_test}
\title{Proc Freq results R data frame for unit test of proc freq function}
\format{Data frame with columns
\describe{
\item{Category}{Iris flowers species}
\item{Frequency}{Frequency of flowers}
\item{Percentage}{Percentage frequency of flowers}
\item{Cumulative.Frequency}{Cumulative frequency}
\item{Cumulative.Percentage}{Cumulative percentage}
}}
\usage{
proc_freq_test
}
\description{
Proc Freq results R data frame for unit test of proc freq function
}
\examples{
data(proc_freq_test)
}
\keyword{datasets}
