% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_proc_freq.R
\docType{data}
\name{results_proc_freq}
\alias{results_proc_freq}
\title{Airline names.}
\format{Data frame with columns
\describe{
\item{carrier}{Two letter abbreviation.}
\item{name}{Full name.}
}}
\source{
<https://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236>
}
\usage{
results_proc_freq
}
\description{
Look up airline names from their carrier codes.
}
\examples{
data(results_proc_freq)
}
\keyword{datasets}
