% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanic_testing.R
\docType{data}
\name{titanic_testing}
\alias{titanic_testing}
\title{Titanic testing data set}
\format{Data frame of 418 observations and 11 features:
\describe{
\item{PassengerId}{id of the passenger}
\item{Pclass}{passenger class on the boat}
\item{Name}{name of the passenger}
\item{Sex}{male / female}
\item{Age}{age of the passenger}
\item{SibSp}{number of siblings/spouses aboard}
\item{Parch}{number of parents/children aboard}
\item{Ticket}{ticket no}
\item{Fare}{price of the ticket}
\item{Cabin}{location of the cabin on the boat}
\item{Embarked}{harbor city of boarding}
}}
\source{
Kaggle Titanic Competition: \url{https://www.kaggle.com/c/titanic/data}
}
\usage{
titanic_testing
}
\description{
Titanic testing data set
}
\examples{
data(titanic_testing)
}
\keyword{datasets}
