% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanic_train.R
\docType{data}
\name{titanic_train}
\alias{titanic_train}
\title{Titanic training data set from Kaggle Competition}
\format{Data frame with 12 columns}
\usage{
titanic_train
}
\description{
Titanic training data set from Kaggle Competition
}
\examples{
data(titanic_train)
}
\keyword{datasets}
