% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanic_validation.R
\docType{data}
\name{titanic_validation}
\alias{titanic_validation}
\title{Titanic validation data set}
\format{Data frame of 179 observations and 12 features}
\source{
Kaggle Titanic Competition: \url{https://www.kaggle.com/c/titanic/data}
}
\usage{
titanic_validation
}
\description{
Titanic validation data set
}
\examples{
data(titanic_validation)
}
\keyword{datasets}
