% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows_to_linux_path.R
\name{windows_to_linux_path}
\alias{windows_to_linux_path}
\title{Change the path windows ==> linux}
\usage{
windows_to_linux_path()
}
\description{
This function allows to change "\" into "/" from system paths
}
