% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_countries_pop.R
\docType{data}
\name{world_countries_pop}
\alias{world_countries_pop}
\title{World countries population from 1960 to 2018}
\format{R Data frame with 15576 obervations and 4 variables
\describe{
\item{country_name}{name of the country}
\item{country_code}{code of the country in three letters}
\item{year}{year from 1960 to 2018}
\item{population}{number of people}
}}
\usage{
world_countries_pop
}
\description{
World countries population from 1960 to 2018
}
\examples{
data(world_countries_pop)
}
\keyword{datasets}
