% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_inertia_ahc.R
\name{clust_centroid}
\alias{clust_centroid}
\title{Centroid of a cluster}
\usage{
clust_centroid(i, data, cluster)
}
\arguments{
\item{i}{integer that represents the cluster number}

\item{data}{R data frame (all columns are required to be numeric types)}

\item{cluster}{character column of the data frame representing the clusters}
}
\description{
This function allows to compute the centroid of a cluster in a R data frame
}
\references{
Link to the author's github repository:
\url{https://www.github.com/Redcart}
}
\author{
Simon CORDE
}
\keyword{centroids}
\keyword{data}
\keyword{frame}
