% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_inertia_ahc.R
\name{compute_inertia}
\alias{compute_inertia}
\title{Inertia of a data frame}
\usage{
compute_inertia(data)
}
\arguments{
\item{data}{R data frame (all columns are required to be numeric types)}
}
\value{
numeric value representing the total inertia
}
\description{
This function allows to compute the inertia of a R data frame
}
\examples{
compute_inertia(mtcars)
}
\references{
Link to the author's github package repository:
\url{https://www.github.com/Redcart/helda}
}
\author{
Simon CORDE
}
\keyword{data}
\keyword{frame}
\keyword{inertia}
