% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_inertia_ahc.R
\name{compute_inertia_ahc}
\alias{compute_inertia_ahc}
\title{Intragroup inertia for choosing best number of cluster in AHC}
\usage{
compute_inertia_ahc(data, max_clusters = 10)
}
\arguments{
\item{data}{R data frame (all columns are required to be numeric types)}

\item{max_clusters}{maximal number of clusters for which we compute intra group inertia}
}
\description{
This function allows to compute the intra group inertia from agglomerative clustering for different number of clusters
}
\references{
Link to the author's github repository:
\url{https://www.github.com/Redcart}
}
\author{
Simon CORDE
}
\keyword{data}
\keyword{frame}
\keyword{inertia}
