% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_calendar_day.R
\name{create_calendar_day}
\alias{create_calendar_day}
\title{Complete empty calendar}
\usage{
create_calendar_day(data, key_variable, time_variable, start_year, end_year)
}
\arguments{
\item{data}{R data frame}

\item{key_variable}{variable name that refers to the key variable in the panel data (ID, ...)}

\item{time_variable}{time variable name that permits to sort observation on a time scale}

\item{start_year}{starting year of the time serie}

\item{end_year}{ending year of the time serie}
}
\value{
a R data frame with the key and time variable for a complete calendar
}
\description{
This function allows to create a complete empty calendar on a year scale
}
\references{
Link to the author's github repository:
\url{https://www.github.com/Redcart}
}
\author{
Simon CORDE
}
\keyword{calendar}
\keyword{fill}
\keyword{gaps}
\keyword{series}
\keyword{time}
