% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_formula.R
\name{create_formula}
\alias{create_formula}
\title{Create a formula}
\usage{
create_formula(data, position = 1)
}
\arguments{
\item{data}{R data frame}

\item{position}{number of the column in the data frame that we want to predict}
}
\value{
string that contains the formula
}
\description{
This function allows to create a formula from the columns of a data frame very quickly
}
\details{
The formula is diplayed with the following format: Y ~ X1 + X2
}
\examples{
data <- iris
str(data)
result_array <- proc_freq(data$Species)

}
\references{
Link to the author's github repository:
\url{https://www.github.com/Redcart}
}
\author{
Simon CORDE
}
\keyword{data}
\keyword{formula}
\keyword{frame}
