% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_formula.R
\name{create_formula}
\alias{create_formula}
\title{Create a formula}
\usage{
create_formula(data, position = 1)
}
\arguments{
\item{variable}{vector on which we want to apply the function}

\item{digits}{integer that specifies the number of decimals we want to keep in the rounded figures}
}
\value{
data frame of dimension [nb of categories x 5].
}
\description{
This function permits to reproduce the output of the SAS proc freq
}
\details{
The five columns diplay the following information:
\itemize{
 \item the first one shows the different categories of the variable on which we want to perform proc_freq
 \item then we have frequency, percentage, cumulative frequency and cumulative percentage
 }
}
\examples{
data <- iris
str(data)
result_array <- proc_freq(data$Species)

}
\references{
Link to the author's github repository:
\url{https://www.github.com/Redcart}
}
\author{
Simon CORDE
}
\keyword{SAS}
\keyword{categorical}
\keyword{frequency}
\keyword{proc_freq}
\keyword{table}
\keyword{variable}
