% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_formula.R
\name{create_formula}
\alias{create_formula}
\title{Create a formula}
\usage{
create_formula(data, position = 1)
}
\arguments{
\item{data}{a R data frame.}

\item{position}{integer representing the position of the column in the data frame that we want to predict.
The other columns are all considered as explanatory variables.}
}
\value{
a string that contains the formula.
The formula is diplayed with the following format: "Y ~ X1 + X2 + ..."
}
\description{
This function allows to create a formula from the columns of a data frame very quickly
}
\examples{
data <- iris
str(data)
result <- create_formula(data = data, position = 4)
result
}
\references{
Link to the author's github package repository:
\url{https://www.github.com/Redcart/helda}
}
\author{
Simon CORDE
}
\keyword{data}
\keyword{formula}
\keyword{frame}
