% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_formula_test.R
\docType{data}
\name{create_formula_test}
\alias{create_formula_test}
\title{String for unit test of create_formula function}
\format{Formula with features of Iris data set}
\source{
<https://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236>
}
\usage{
create_formula_test
}
\description{
String for unit test of create_formula function
}
\examples{
data(create_formula_test)
}
\keyword{datasets}
