% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_to_fill.R
\name{gap_to_fill}
\alias{gap_to_fill}
\title{Filling intermdediate gaps in a time serie}
\usage{
gap_to_fill(data, gap_variable, key_variable, time_variable, digits = 2)
}
\arguments{
\item{data}{R data frame}

\item{gap_variable}{name of the variable we want to fill the start and end gaps}

\item{key_variable}{variable name that refers to the key variable in the panel data (ID, ...)}

\item{time_variable}{time variable name that permits to sort observation on a time scale}

\item{digits}{number of decimals to keep for the rounding (by default 2)}
}
\value{
a R data frame of dimension containing the original columns and new ones
}
\description{
This function allows to fill intermediate gaps in panel data by linear interpolation
}
\details{
The news columns are:
\itemize{
 \item _corrected_2: the variable with intermediate gaps filled
 }
}
\references{
Link to the author's github repository:
\url{https://www.github.com/Redcart}
}
\author{
Simon CORDE
}
\keyword{fill}
\keyword{gaps}
\keyword{interpolation}
\keyword{series}
\keyword{time}
