% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_procedure.R
\name{kmeans_procedure}
\alias{kmeans_procedure}
\title{Kmeans procedure}
\usage{
kmeans_procedure(
  data,
  columns,
  threshold_min,
  threshold_max,
  verbose = FALSE,
  seed = 42
)
}
\arguments{
\item{data}{R data frame}

\item{columns}{vector of columns names of the data frame on which we perform the kmeans algorithm. These features have to be numeric.}

\item{threshold_min}{integer minimum size for cluster.}

\item{threshold_max}{integer maximum size fo cluster.}

\item{verbose}{if set to TRUE print the current state of the procedure (by default set to FALSE).}

\item{seed}{seed for the random call (if we want the output to be reproducible).}
}
\value{
R data frame
}
\description{
This function allows to perform kmeans clustering with some desired output sizes for clusters
}
\details{
The R data frame contains the id of the original data frame and a column `cluster` representing
the cluster to which the observation belongs to.
}
\references{
Link to the author's github repository:
\url{https://www.github.com/Redcart}
}
\author{
Simon CORDE
}
\keyword{cluster}
\keyword{kmeans}
\keyword{sizes}
