% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_freq.R
\name{proc_freq}
\alias{proc_freq}
\title{SAS proc freq in R}
\usage{
proc_freq(variable, digits = 4)
}
\arguments{
\item{variable}{vector on which we want to apply the function.}

\item{digits}{integer that specifies the number of decimals we want to keep in the rounded figures.}
}
\value{
a R data frame of dimension [number of categories x 5].
The five columns display the following information:
\itemize{
 \item Category: different categories of the original categorical variable
 \item Frequency
 \item Percentage
 \item Cumulative.Frequency
 \item Cumulative.Percentage
 }
}
\description{
This function permits to reproduce the output of the SAS proc freq
}
\examples{
data <- iris
str(data)
result <- proc_freq(data$Species)
result
}
\references{
Link to the author's github package repository:
\url{https://github.com/Redcart/helda}
}
\author{
Simon CORDE
}
\keyword{SAS}
\keyword{categorical}
\keyword{frequency}
\keyword{proc_freq}
\keyword{table}
\keyword{variable}
