% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_end_to_fill.R
\name{start_end_to_fill}
\alias{start_end_to_fill}
\title{Function for filling start and end of time series}
\usage{
start_end_to_fill(
  data,
  calendar,
  gap_variable,
  key_variable,
  time_variable,
  digits = 2
)
}
\arguments{
\item{data}{R data frame}

\item{calendar}{complete empty calendar (performded by \code{create_calendar_day})}

\item{gap_variable}{name of the variable we want to fill the gaps}

\item{key_variable}{variable name that refers to the key variable in the panel (ID, ...)}

\item{time_variable}{time variable name that permits to sort observation on a time scale}

\item{digits}{number of decimals to keep for the rounding}
}
\value{
a R data frame of dimension containing the original columns and new ones
}
\description{
This function allows to fill the start and end gaps of time series by doing repetition
}
\details{
The five columns diplay the following information:
\itemize{
 \item the first one shows the different categories of the variable on which we want to perform proc_freq
 \item then we have frequency, percentage, cumulative frequency and cumulative percentage
 }
}
\references{
Link to the author's github repository:
\url{https://www.github.com/Redcart}
}
\author{
Simon CORDE
}
\keyword{fill}
\keyword{gaps}
\keyword{series}
\keyword{time}
