% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_end_to_fill.R
\name{start_end_to_fill}
\alias{start_end_to_fill}
\title{Function for filling start and end gaps in time series}
\usage{
start_end_to_fill(data, calendar, gap_variable, key_variable, time_variable)
}
\arguments{
\item{data}{R data frame}

\item{calendar}{R data frame complete empty calendar (as one can perform with \code{create_calendar_day})}

\item{gap_variable}{character that represents name of the variable we want to fill the start and end gaps}

\item{key_variable}{character that represents variable name that refers to the key variable in the panel data (ID, ...)}

\item{time_variable}{character that represents time variable name that permits to sort observation on a time scale}
}
\value{
a R data frame of dimension containing the original columns and a new one
}
\description{
This function allows to fill the start and end gaps of a time series
by doing repetition of next (for the start) and previous values (for the end)
}
\details{
The new column is:
\itemize{
 \item _corrected_1: the variable with starts and ends filled
 }
}
\references{
Link to the author's github package repository:
\url{https://www.github.com/Redcart/helda}
}
\author{
Simon CORDE
}
\keyword{fill}
\keyword{gaps}
\keyword{series}
\keyword{time}
