% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanic_testing.R
\docType{data}
\name{titanic_testing}
\alias{titanic_testing}
\title{Titanic testing data set}
\format{Data frame of 418 observations and 10 features}
\source{
Kaggle Titanic Competition: \url{https://www.kaggle.com/c/titanic/data}
}
\usage{
titanic_testing
}
\description{
Titanic testing data set
}
\examples{
data(titanic_testing)
}
\keyword{datasets}
