% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanic_training.R
\docType{data}
\name{titanic_training}
\alias{titanic_training}
\title{Titanic training data set}
\format{Data frame of 712 observations and 12 features}
\usage{
titanic_training
}
\description{
Titanic training data set
}
\examples{
data(titanic_training)
}
\keyword{datasets}
