% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows_to_linux_path.R
\name{windows_to_linux_path}
\alias{windows_to_linux_path}
\title{Convert windows path into linux path}
\usage{
windows_to_linux_path()
}
\description{
This function allows to make conversion of windows path into linux path
}
\details{
When the function is called, a prompt asks for the windows path to be converted
in the R console. Enter a windows path or copy paste one. Then type ENTER.
The Linux converted path appears.
}
\references{
Link to the author's github repository:
\url{https://www.github.com/Redcart}
}
\author{
Simon CORDE
}
\keyword{SAS}
\keyword{categorical}
\keyword{frequency}
\keyword{proc_freq}
\keyword{table}
\keyword{variable}
